Changed `Sorter::count()` to return `size_t` instead of `int`.
Also applies to `DRFSorter::count()`. The implementations return
`hashmap::count()` which is a `size_t`. Converting `size_t` to `int`
implicitly generates a "possible loss of data" warning.

Review: https://reviews.apache.org/r/64730

+++ src/master/allocator/sorter/drf/sorter.cpp
@@ -584,7 +584,7 @@ bool DRFSorter::contains(const string& clientPath) const
 }
 
 
-int DRFSorter::count() const
+size_t DRFSorter::count() const
 {
   return clients.size();
 }
